+++ prow/BUILD
@@ -35,8 +35,8 @@ filegroup(
         "//prow/hook:all-srcs",
         "//prow/jenkins:all-srcs",
         "//prow/kube:all-srcs",
-        "//prow/npj:all-srcs",
         "//prow/phony:all-srcs",
+        "//prow/pjutil:all-srcs",
         "//prow/plank:all-srcs",
         "//prow/plugins:all-srcs",
         "//prow/report:all-srcs",
