+++ prow/cmd/deck/BUILD
@@ -35,7 +35,7 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/jenkins:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
         "//vendor/github.com/NYTimes/gziphandler:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
