+++ prow/cmd/mkpj/BUILD
@@ -19,7 +19,7 @@ go_library(
     deps = [
         "//prow/config:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
     ],
