+++ prow/jenkins/BUILD
@@ -17,7 +17,7 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
         "//prow/report:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
     ],
@@ -44,6 +44,6 @@ go_test(
     deps = [
         "//prow/config:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
     ],
 )
