+++ prow/pjutil/pjutil.go
@@ -14,8 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package npj contains helpers for creating ProwJobs.
-package npj
+// Package pjutil contains helpers for working with ProwJobs.
+package pjutil
 
 import (
 	"fmt"
