+++ prow/pjutil/pjutil_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package npj
+package pjutil
 
 import (
 	"reflect"
