+++ prow/plank/BUILD
@@ -13,7 +13,7 @@ go_test(
     deps = [
         "//prow/config:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
     ],
 )
 
@@ -24,7 +24,7 @@ go_library(
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//prow/kube:go_default_library",
-        "//prow/npj:go_default_library",
+        "//prow/pjutil:go_default_library",
         "//prow/report:go_default_library",
         "//vendor/github.com/bwmarrin/snowflake:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
